requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
groq
spacy
textstat
